# docs/assets/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/labeling/projects": {
      "get": {
        "operationId": "LabelingController_getProjects",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by creator",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of labeling projects with their field schemas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelingProjectResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get all labeling projects",
        "tags": [
          "labeling"
        ]
      },
      "post": {
        "operationId": "LabelingController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Newly created labeling project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelingProjectResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Create a new labeling project",
        "tags": [
          "labeling"
        ]
      }
    },
    "/api/labeling/projects/{id}": {
      "get": {
        "operationId": "LabelingController_getProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Labeling project with full field schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelingProjectResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get project details with field schema",
        "tags": [
          "labeling"
        ]
      },
      "put": {
        "operationId": "LabelingController_updateProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated labeling project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelingProjectResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Update project",
        "tags": [
          "labeling"
        ]
      },
      "delete": {
        "operationId": "LabelingController_deleteProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Delete project and all associated data",
        "tags": [
          "labeling"
        ]
      }
    },
    "/api/labeling/projects/{id}/fields": {
      "get": {
        "operationId": "LabelingController_getFieldSchema",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordered list of field definitions for the project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FieldDefinitionResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get field schema for project",
        "tags": [
          "labeling"
        ]
      },
      "post": {
        "operationId": "LabelingController_addField",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFieldDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Newly created field definition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldDefinitionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Add a field to the project schema",
        "tags": [
          "labeling"
        ]
      }
    },
    "/api/labeling/projects/{id}/fields/{fieldId}": {
      "put": {
        "operationId": "LabelingController_updateField",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fieldId",
            "required": true,
            "in": "path",
            "description": "Field ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFieldDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated field definition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldDefinitionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Update a field definition",
        "tags": [
          "labeling"
        ]
      },
      "delete": {
        "operationId": "LabelingController_deleteField",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fieldId",
            "required": true,
            "in": "path",
            "description": "Field ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Field deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Delete a field from schema",
        "tags": [
          "labeling"
        ]
      }
    },
    "/api/labeling/projects/{id}/documents": {
      "get": {
        "operationId": "LabelingController_getProjectDocuments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of labeled documents with their labels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabeledDocumentResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get all documents in project",
        "tags": [
          "labeling"
        ]
      },
      "post": {
        "operationId": "LabelingController_addDocumentToProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document added to the project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabeledDocumentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Add a document to project",
        "tags": [
          "labeling"
        ]
      }
    },
    "/api/labeling/projects/{id}/upload": {
      "post": {
        "operationId": "LabelingController_uploadLabelingDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelingUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document uploaded and queued for OCR processing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadLabelingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Upload a document into a labeling project",
        "tags": [
          "labeling"
        ]
      }
    },
    "/api/labeling/projects/{id}/documents/{docId}": {
      "get": {
        "operationId": "LabelingController_getProjectDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "description": "Document ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Labeled document with all its labels and underlying document data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabeledDocumentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get document with labels",
        "tags": [
          "labeling"
        ]
      },
      "delete": {
        "operationId": "LabelingController_removeDocumentFromProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "description": "Document ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document removed from project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDocumentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Remove document from project",
        "tags": [
          "labeling"
        ]
      }
    },
    "/api/labeling/projects/{id}/documents/{docId}/download": {
      "get": {
        "operationId": "LabelingController_downloadLabelingDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "description": "Labeling Document ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary file content (PDF or image)"
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Download a labeling document file",
        "tags": [
          "labeling"
        ]
      }
    },
...
    "/api/labeling/projects/{id}/documents/{docId}/labels/{labelId}": {
      "delete": {
        "operationId": "LabelingController_deleteLabel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "description": "Document ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": true,
            "in": "path",
            "description": "Label ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Label deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Delete a specific label",
        "tags": [
          "labeling"
        ]
      }
    },
    "/api/labeling/projects/{id}/documents/{docId}/ocr": {
      "get": {
        "operationId": "LabelingController_getDocumentOcr",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "description": "Document ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw OCR result from Azure Document Intelligence"
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get OCR data for document",
        "tags": [
          "labeling"
        ]
      }
    },
...
  },
  "info": {
    "title": "Document Intelligence Platform API",
    "description": "Comprehensive REST API for document processing, OCR, workflow orchestration, model training, and human-in-the-loop review",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "keycloak-sso": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "LabelingProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "field_schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldDefinitionResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "created_by",
          "created_at",
          "updated_at",
          "field_schema"
        ]
      },
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Project name"
          },
          "description": {
            "type": "string",
            "description": "Project description"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}